// src/types.ts
export interface IGenerateOptions {
  topType?: string[];
  accessoriesType?: string[];
  facialHairType?: string[];
  facialHairColor?: string[];
  clotheType?: string[];
  eyeType?: string[];
  eyebrowType?: string[];
  mouthType?: string[];
  skinColor?: string[];
  hairColorTypes?: string[];
  hatColor?: string[];
  clotheColor?: string[];
}